requirements.txt
Pillow
piexif
google-cloud-vision
kindwise-api-client
paho-mqtt
flask